WIP: torrent item ui tweak
